Bug #34229520 : Ndb MTA unordered commits with log_replica_updates=1
Problem
-------

Commit ordering is not preserved for NDB transactions, even if
replica_preserve_commit_order is turned on.

Analysis / Root-cause analysis
------------------------------

Commit ordering is implemented in several places:

1. binlog.cc, in the `ordered_commit` function
2. handler.cc, in the `ha_commit_low` function

Ordering is enabled in case binlogging is turned on and binlog has
transaction information in its local caches or binlogging option is
disabled. NDB uses MYSQL_BINLOG as transaction coordinator however
it does not cache data locally. Therefore, commit ordering is never
taken into account.

Solution
--------

Ordering condition in the `ha_commit_low` function is extended to
take into account engines that do not use binlog cache manager -
newly introduced flag 'is_low_level_ordering_enabled' is checked.
The flag is set up in the 'binlog::commit' function to true in
case binlog caches are disabled or empty and binlogging is
enabled in the current call to the function. This condition is
checked by the implemented
'is_current_stmt_binlog_enabled_and_caches_empty' function
which may be safely called in case binlogging is disabled
or binlog cache manager is not initialized. The check is done
in the 'binlog::commit' because there are cases in which binlog
caches were emptied after the thread entered the ordered_commit
function in the MYSQL_BIN_LOG.
Please also note that it is important to check
opt_bin_log in is_current_stmt_binlog_log_replica_updates_disabled,
because of statements such as ALTER TABLE OPTIMIZE PARTITION,
where the last call to trans_commit_stmt in the
mysql_inplace_alter_table (Implicit_substatement_guard disabled)
is not the last call.

Signed-off-by: Karolina Szczepankiewicz <karolina.szczepankiewicz@oracle.com>
Change-Id: I942f0a42aa9b3a033279a58deb8d37345b887b90