Bug#31072098: Assertion ".var->check.thd"
Patch developed by Sergey Glukhov.

The problem is not related to the hints itself but rather to
mysqld startup option behaviour. When startup option is used
with the 'maximum' prefix, upper bound is set but current
value(default value) is not adjusted according to the new
limit and can be bigger that 'maximum' value.

In the case of SET_VAR hints, we restore original value
at the end of the query execution. In TRADITIONAL mode it
fails with an error (since original value is bigger than
upper bound) which leads to assertion failure.

The fix is to adjust current option value if it is bigger
with user defined maximum value.

Change-Id: I57df7d3cb0c54c10bf3f7201f4d3959e9b729587