Bug#34341503: TPC-DS Q18 SF1 result diff old optimizer vs. hypergraph
The hypergraph optimizer sometimes gives wrong results when queries with
GROUP BY ... WITH ROLLUP are passed through a temporary table prior to
aggregation.

The problem occurs due to the failure of an assumption that the base
columns of rollup items are always included in the temporary table.
When base columns from the GROUP BY list are left out of the temporary
table, the replacement logic fails to update rollup items to point
into the temporary table, and the AggregateIterator's logic for
reading and restoring columns during the execution of GROUP BY fails.

This shows up as a bug where the value of the rollup item in a result
row is off by one row because the underlying iterator is one step ahead
and has not been restored correctly.

This patch fixes the problem by ensuring that non-rollup descendants of
rollup items are added to the temporary table.

Change-Id: I31ace34296639563fc1935e682213e25e26285e0