Bug#34486874 Error 4829 'FirstInner/Upper has to be an ancestor or a sibling' from NDBCLUSTER
Update of patch set as a result of review comments and later investigation.

It turns out that an extra check of required nest_dependencies being
covered by the m_ancestor_nests is completely dead code. I.e. its conditions
for rejecting a table as pushable was never found.

This was probably required once upon time. As code has matured, and bugs
being fixed, the condition checked in this code section is likely
covered by other checks of dependencies being within the nest-reach
we are allowed to refer. The splitting of m_upper_nests into
a 'static' m_upper_nests and m_ancestor_nests (first patch
in this series) seems to also play a role here. Without that patch
we found MTR test queries which passed the first if-clause in
the removed code, the error return was not hit even here though even
after running > 1M Random Generated Join Queries. Didn't
investigate further when this code became obsolete.

A couple of asserts was added as a guard against required
nest_dependencies not being being covered by the m_ancestor_nests.

Change-Id: I6121515f5634976dcfafd5b4754f1fcd749d5457