bug#34530397 InnoDB: outdated/confusing code in row0mysql.cc
The comment in the row_create_index_for_mysql routine refers to a
que_run_threads call, which is not present in the code anymore.
Also, the local variable table_name really appears to be unnecessary --- in
the current form of the code, it is only used before the index is destroyed,
thus one can simply use index->table_name instead.

Change-Id: Id9ed521494f4a5010115c6dfc196e70fe8b2c8d7

Approved by Marcin Babij <marcin.babij@oracle.com>

Change-Id: I148b21c9f93ee58c8e008f2db6be43e9d3340ee5