Bug#34494877: WL#14449: Offload issue: RapidException (3): rpdrqctr_transcode.c:1447 @ rpdoqc_
When a non-merged derived table is found to always return zero rows,
the hypergraph optimizer will still produce a plan which reads the
contents of the subquery into a temporary table, and the outer query
block will see a table scan on this always empty temporary table
instead of a ZERO_ROWS access path. This makes it more difficult to
see that subtrees of the join can be pruned away in the outer query
block, and can lead to creating actual join access paths in cases
where a ZERO_ROWS access path would suffice.

This patch makes the hypergraph optimizer stop adding a
materialization step on always empty derived tables. This makes it
possible to optimize away more subtrees from the final plan, and in
some cases even optimize away the entire query.

Change-Id: Ie9404f75dad0ef40b4bd462853d87e6b5753041c