Bug#34538065 NDBXFRM1 only write checksum to trailer in append (write_forward) mode.
Fixup 8.0.29: WL#14285 [10] ndbxfrm_file: class for accessing backup and
ndbfs files.

The crc32 was always written to file trailer even if it was not uptodate
or never applicable for the file.

The bug has no impact on old versions since the only encrypted files was
backup files which always was written in an append mode fashion and a
valid crc32 sum was always written to trailer.

Also when reading files the checksum was never checked whether it
existed in trailer or not.

If in future the checksum checks are added when reading file only valid
crc32 sums should be stored, to allow for example encrypted redolog
files to be read with such future ndbxfrm tool it must not have a crc32
sum stored in trailer now since it is wrong and that will cause future
decryption of redolog files with ndbxfrm tool to fail.

Change-Id: I4f8c9b5a2ff24fdc3e53e797edb97f9a1a62414b