Bug#34226552 clone_files_recovery in case of log_downsize_requested depends on BP state
Description
-----------
In the Bug#34485875, we want to remove dead code that was added to
support upgrade scenario from 5.7.7 or before. After we removed that
dead code, it was found that some tests in the clone suites were
failing if run in the address sanitizer.
Tests are : clone.ddl_alter_copy clone.ddl_basic or clone.ddl_truncate

While investigating it was found that this problem was always present
but was masked. Therefore, irrespective we remove the dead code or not
we should fix the masked problem.

In Kuba's words :

```
The explanation for this phenomenon seems to be that this
"unrelated check" brings the page to BP, which would otherwise be empty,
which lets the clone_files_recovery avoid IO of this single page, which
in turn lets it bypass validation, which in turn hides the bug that we
expect the log system to be initialized, while in fact it was just freed
a line before.
```

If we don't recreate the redo log then the unrelated check would not
kick in hence this problem will be addressed. We anyways want to get
rid of the synchronous redo log resizing code.

Fix:
----
- Removed the check for forced synchronous redo downsizing which was
  added to retain the old behavior that is not needed anyways, why ?
  Due to the following reasons :
   + mysql-trunk already makes all of the pages clean before finishing
     recovery
   + This means flush list is empty
   + This means checkpoint lsn can jump till the end of log
   + This means log is logically empty
   + This means file governor thread (that works asynchronously) can:
     remove all  older files (except the last one), truncate the current
     file to end right after logical end's block, and also use the
     "generate dummy log entries" technique to complete the file's last
     block, move to the next file, remove the old file, and resize
     everything to new size.

- Hence unrelated checks are not kicked in and ASAN doesn't complain
  anymore.

- Also learnt that the test to verify the redo log downsizing in case of
  innodb read only mode was fake. If server started in the
  --innodb-read-only-mode then log system is not initialized at all.
  That means we have no way to find out size of the actual active redo
  log size. We only know that only active redo log files remain
  available in the redo only mode and the use specified value of
  --innodb-redo-log-capacity is ignored.  Remove the fake code,instead
  added a warning in the server error logs.

- Found another bug that is if user tries to set the
  innodb-redo-log-capacity when the server started with
  --innodb-read-mode=ON then user connection hangs.
  Created a bug Bug #34537146 for it.

Testing
-------

- Ran the following three tests 50 times each, with santizer ON to
  ensure the actual problem is fixed.
[ 77%] clone.ddl_truncate                       w26 [ pass ]  27464
[ 94%] clone.ddl_basic                          w5  [ pass ]  43646
[ 98%] clone.ddl_alter_copy                     w17 [ pass ]  103082
[ 99%] shutdown_report                          w0  [ pass ]
[100%] sanitize_report                          w0  [ pass ]
------------------------------------------------------------------------
The servers were restarted 0 times
The servers were reinitialized 0 times
Spent 8796.015 of 416 seconds executing testcases

Completed: All 152 tests were successful.

- In scenario#2 in the log_read_only.test file was fake.
  Improved the test to verify that we check for warning message in the
  log file. Suppressed the warning in few tests as needed.

- Executed important innodb related tests successfully.
- Fixed serveral test due to the cascade effect of adding the new
  warnings

Change-Id: Iba808d835afd9a92a5e6cba5b3ffc0c1340d249a