WL#15008: Remove Abstract Query Plan (AQP) interface
Add non-tables to nest table_map as well

Normally a leaf-node in the AccessPath tree will refer a
'BasicTable'. However there are cases where the leaf-node does not
refer a table at all, e.g. ZERO_ROWS, TABLE_VALUE_CONSTRUCTOR, ...
These will now be represented in the table_maps being constructed,
and added to the m_tables[] with a nullptr table member.
(and later checked for not being a nullptr where required)

Note that such non-tables will not be possible to push. However
they are semantically still part of the (outer-)join_nest. Thus
they should be visible as nest members and taken into account
when considering, e.g in semi- and outer-joins where there are
limitations requiring all tables in a nest to be pushed.
(Else MTR tests will fail after next patch in this series...)

Currently we do not add such non-tables to the nest bitmaps.
This patch add these to the 'nests' as well. A couple of
additional checks for 'table == nullptr' had to be added as well.

Change-Id: I2698a9d17af8ca993b9a013df34f872d35729eb1