WL#349 INTERSECT and EXCEPT table operators
Contains code for set operators INTERSECT [DISTINCT | ALL] and EXCEPT
[All | DISTINCT ] performing matching (deduplication) by means of a
materialized internal tmp table: it is indexed by a hash of the row
contants.

This is the same mechanism used for mixed UNION ALL / UNION DISTINCT
which is not at the top level; in such a case the intermediate
temporary table does not have a primary key consisting of the entire
row, but falls back on a non-unique hash hey instead.

While this solution is workable, it is likely that leveraging
the MySQL hash join machinery could improve INTERSECT, EXCEPT
performance, but this is not attempted here.

Change-Id: I5300584c5180a0f799bbe7b1973571267316cf86