WL#15008: Remove Abstract Query Plan (AQP) interface
Cleanup / unify interface:

The AQP interface had the isInnerJoined() and isOuterJoined()
methods, while the pushed_table object had the is_semi_joined()
and is_anti_joined() method for checking if a table was
semi-joined with an ancestor.

As the former AQP methods now are members of pushed_table
as well, these need to be unified.

Patch replace is_semi_joined() and is_anti_joined() with resp.
isSemiJoined() and isAntiJoined(). These are implemented in
similar (and simpler) way as the existing isOuterJoined()
and isSemiJoined(), comparing the ancestor tab_no with
the 'first' in the existing nest

Builds on the previous patch, introducing these 'first'
members.

Also change some if-checks where we used some less obvious logic
to check for the above join-types to use these methods.

Change-Id: I6ee0efd97b8cfacebb7415ae7bf8da79ec4562ad