Bug#34523627: Mysql-server crashed with Assertion `false' failed
When creating natural join columns, a hidden field added as part of
a materialized derived table is used to make the join condition.
This condition is later used to check if its eligible for pushdown
to derived tables. While checking, it fails to retrieve the field
from derived table since its a hidden field. Even when condition
pushdown optimization is off, query fails for "EXPLAIN".
Solution is to reject all hidden columns added internally.
Earlier the check was done only for hidden columns added for
functional indexes.

Change-Id: I94eb49b3988816ccf91f9a494e41579454e11f9d