Bug#34506124 - Replace manual checkpoint actions in srv_start() with log_make_latest_checkpoint
Problem : The bug description is detailed in
https://mybug[.]mysql[.]oraclecorp[.]com/orabugs/site/bug[.]php?id=34506124.
In essence, during initialization, checkpoint_lsn can become smaller than
flushed_lsn in srv_start() when create_new_db is true. This happens because
flushed_lsn header is being written to disk while the checkpointer thread is stopped.
This condition can lead to warning ER_IB_MSG_RECOVERY_CHECKPOINT_FROM_BEFORE_CLEAN_SHUTDOWN.
Currently the lsn values are subsequently equal during slow shutdown of the server.
This is why the warning is not getting caught today.

Fix : On inspection, it is seen that -
   The mentioned code is there to write some value to flushed_lsn header
   But this value is immediatelly overwritten by dblwr::v1::create() back to zero
   And if we crash before this overwrite, we can't boot anyway because mysqld.ibd is missing
   And if we crash afterwards we have 0 in the header
   And the value is wrong to begin with as it doesn't correspond to old nor new checkpoint lsn at all

   This is why it's best to remove this section of the code.

Change-Id: I4ff4801034137995fc88cf456af50f72b3c77370