Bug#34548367 Invalid error reaches client where max connections reached on Windows.
Whenever the max connections limit is reached on the Router side while the Router
accepts the new user connection, it sends back an error and closes the connection.
The problem is that the close is not graceful, which  sometimes prevents the error
from reaching the client before the connection is dropped. That results is the
client getting an error like 'MySQL server has gone away' instead of the expected
'Too many connections to MySQL Router'.

This patch replaces the socket close with ae graceful shutdown, giving the
client a chance to read the error before the connection is dropped.

Change-Id: I4a731f7896c38fd565ffced7db44d9c09b73208e