Bug#34556068 Encrypted backups can not be restored on system with other endian
Encrypted backups can not be restored on system with other endian.

For example backups taken on a little endian (i.e. x86) system can not
be restored on a big endian (sparc) system, and vice versa.

Encrypted backups uses the NDBXFRM1 file format for encrypting the
backup files.

The NDBXFRM1 headers and trailers are written in the systems native
endian.

When reading NDBXFRM1 file the reader should detect whether other endian
have been used and if so it should byte swap affected parts of header
and trailer.

There is a bug in the code for reading NDBXFRM1 files that if file is
not using the native endian the reader fail to detect the endianness.

Instead it will assume file is not encrypted or even a NDBXFRM1 file.

This in turn will cause the restore code to fail since it can not find
backup data magic ("NDBBCKUP").

Support for encrypted backups is added in test
ndb.ndb_print_backup_file.

Check that ndbxfrm tool can handle both little and big endian is added
in ndb.ndb_secretsfile_reader.

Encrypted backups generated on little endian and big endian systems are
recorded and restore of them is tested in new test
ndb.ndb_restore_encrypted_backups.

This last test is disabled due to
Bug#11761083 NDB : MYSQLD DEFAULT VALUES IN FRM EMBEDDED IN BACKUP NOT ENDIAN-CONVERTE.

Change-Id: Ib3f282d657329517b0bf9593230d99ad3ea1ca91