8.0.30 merge - fix archive log decryption issues
The issues is caused by use 'auto' keyword which did a copy of the encryption
object. The reference to copied object is stored elsewhere and the newly
created objected is destroyed.

Thus we are using wrong encryption key and iv causing decryption uses

Fix:
----
Take reference to the log_sys encryption info instead of a copy.