Bug#34506098 log_files_generate_dummy_records() might write before checkpoints are allowed
log_files_generate_dummy_records method validates for
log_checkpointer_is_active to ensure that older log files
can be reclaimed. But this check alone is insufficient
as it can be true in early stages of server startup,
where checkpoints are still not allowed yet (dictated by
m_allow_checkpoints flag).

In addition, the invocation of
log_files_generate_dummy_records which is dependent
result of log_files_filling_oldest_file_takes_too_long
should be restricted based on the condition that both
the checkpointer thread  is active and checkpoints
are enabled so that we generate dummy records
only when we are sure the older REDO log file can
be reclaimed.

Also, the method log_files_consuming_oldest_file_takes_too_long
requires the same checkpoint related checks as it
decides if the checkpointer should be rushed to consume
REDO logs, which is not needed if checkpoints are
not allowed in the first place.

The patch addresses the above 3 issues by adding
both the checks in the above mentioned methods.

Change-Id: Ia0fed6baaf6a15c074ef6b7b9567459a79932f7d