8.0.30 merge - stabilize test case
Log block missmatch issue.
PXB always try to scan redo logs using RECV_SCAN_SIZE, which is 64K on
an instance witih 16K page.
When we are reading the last available block, there is a chance that
only part of the 64K read size will have a recent block (which has
OS_FILE_LOG_BLOCK_SIZE size). Meaning that part of the buffer could
contain data from previous scan.

As part of the 8.0.30 merge, we wrongly adjusted this case to only
consider the scan completed if the read block was 0, which is wrong and
we should consider the scan completed if block checksum is ok and
the nro of new block is lower than the one we expect.