BUG#34549470: setting replication_optimize_for_static_plugin_config=ON can cause MySQL cores
MySQL plugins can be register to notifications about server and
transactions life-cycle, allowing them to handle those events.
These notifications are delivered through hooks which do walk
through the registered plugins, calling the listeners.
One of the stubs of these hooks was not considering observers that
only didn't implement all listeners, which was causing a signal 11.

To solve the above issue, the stub implementation now does consider
that not all listeners need to be implemented.

Change-Id: I9c8adb9b4c7d7be41081b813b62d1c8c58df2b80