Bug #34583577 Remove dead code identified by gcov
Testing of the patch for:
Bug #34570345 Update fastcov.cmake to the latest fastcov.py
revealed some dead code. Remove it.

Most of the code has been used in earlier versions of MySQL,
but subsequent refactorings/reimplementations have failed
to remove old stuff.

The patch also adds a few annotations to unused code, to show how to
make gcov ignore it.

Change-Id: Icd41241ff0d06cc0337d7e8d667610c9a98d9ce2