Bug#34588203: Cleanups in unittest/gunit/optimizer_test.h
Make OptimizerTestBase an ordinary class, not a class template. It was
a template in order to allow it to be used as the base class of both
parameterized tests (which inherit from TestWithParam) and
non-parameterized tests (which inherit from Test). But it is fine that
both types of tests inherit from Test. Parameterized tests just have
to inherit from WithParamInterface in addition to Test.

Add include guards to the header file.

Change the "static" functions to "inline" functions in the header
file. The style guide says internal linkage should not be used in
header files.

Run IWYU to make sure it contains the correct #include statements.

Change-Id: I3fcc35e9c350aa9994cae47e1a0a9397fbd20a15