Bug#34569676 Static routing accepting port never reopened when port busy after quarantine.
When all the routing potential destinations become unavailable, the Router
shuts down the accepting port. The destinations are put into the quarantine
and whenever one of them becomes available again the accepting port is
reopened. The problem is that if that fails (for example because
the accepting port is been taken by another application etc.), the next
reopening attempt is only triggered by the metadata cache refresh.
That works only for the InnoDB Cluster destinations not for the ones
that are statically defined in the configuration file.

This patch adds a retry mechanism (every 1 second) when reopening the
accepting port for standalone destinations fails.

Change-Id: I5c573c7ffc402dd38af2f5b6b5f7df84001a0420