Bug#34463842: Result diff seen with hypergraph off and on
Wrong results were seen in some queries using GROUP BY and where a
portion of the HAVING clause could be optimized away.

The wrong result was caused by UpdateReferencesToMaterializedItems()
not correctly replacing expressions in the HAVING clause with
references to the corresponding materialized result in a temporary
table. The HAVING clause therefore read values from the base tables's
row buffers instead of the materialized results.

It is fixed by making UpdateReferencesToMaterializedItems() call
FindReplacementOrReplaceMaterializedItems() with the need_exact_match
flag set to false when called on HAVING clauses, allowing it to see
through the Item_aggregate_ref and correctly replacing the expression
with a reference to a temporary table column.

Change-Id: I5d6227c33cd76adf2200fac69ae3cf232d314032