Bug#33294005: mysql server crash with fulltext index search [3/3, rollup]
Some grouped queries that returned full-text search results could make
the server exit. This happened if the query execution plan used
streaming aggregation instead of temporary table aggregation and
failed to materialize the full-text search results before aggregation.

This patch handles the bug for GROUP BY WITH ROLLUP. In this case,
streaming aggregation is currently the only aggregation method that
can be used, so we cannot force temporary table aggregation, like we
did for the explicitly grouped non-rollup queries in the previous
patch. Adding a materialization step before the aggregation, like the
hypergraph optimizer already does, would solve it. But that is not yet
implemented in the old optimizer.

For now, as a stopgap, if a query is using GROUP BY WITH ROLLUP and
the chosen execution plan attempts to access non-aggregated results
from the full-text search MATCH function after aggregation, an error
is raised to prevent the query from returning wrong results or make
the server exit.

Change-Id: Iec7d9b8e8ad2d14db9c323ac528b26a07caead60