Bug#34619621 Router sporadically crashes on shutdown if client hit max_connect_errors limit.
The crash sometimes happens (the reproduction rate is quite low) when the
Router shutdown is preceeded by the client reaching the max_connect_errors.
In such scenario the Router sends error to the client and postpones the
client socket shutdown to the async operation. When this async operation
is still not executed when the Router shuts down it leads to the
scenario where the io_context object is destroyed which leads to that
AsyncOp destruction. The AsyncOp keeps the Socket object alive with a
smart pointer so the Socket is been destructed calling close() method.
That method tries to unregister the socket from the io_context object
which is partially destructed.

This patch changes the io_context destructor in a way that it clears
the internal asyn_ops map before the AsyncOps objects themselves are
destructed. This prevents the destructors of AsyncOps from potentially
using that internal map later once it is half-destroyed.

Change-Id: I37f195d151abf1e35212b16102871c93386175b7