Bug#34572136 Heap-use-after-free (my_mb_wc_utf8_prototype at strings/mb_wc.h:115)
We have an UPDATE statement with a where clause containing a window
function (LAST_VALUE). In the select list, there is an EXISTS subquery
which references a blob column (an argument to a LIKE function) in the
table we are updating.

In BufferingWindowIterator::Read, there is a section where we call
copy_funcs(... , CFT_HAS_NO_WF) for non-window functions, cf. this
comment in the code:

  /*
    This saves the values of non-WF functions for the row. For
    example, 1+t.a. But also 1+LEAD. Even though at this point we lack
    data to compute LEAD; the saved value is thus incorrect; later,
    when the row is fully computable, we will re-evaluate the
    CFT_NON_WF to get a correct value for 1+LEAD.
  */

This is called for the LIKE function which has as its first argument
an outer reference (a REF item) to the blob.

The slice setting at this point is to the output table from the window
function.  This leads the ref item mentioned to read from the *output*
table, which is wrong, since the fields are not there yet (only after
copy_fields): the blob should reference the *input* table, as it does,
btw, for the first item in the select list in the repro case. But
since for the LIKE argument in the subquery, this is an outer
reference, we access the column via a REF item, which unfortunately
doesn't point to the input table at this point, so we read garbage,
i.e. we try to interpret the output table's record as containing the
blob.

Note: removing rows from the test case or removing other parts of the
query causes the failure to go away, hence the number of rows included
in the test.

Fix: Make sure the current slice is the input table before we call
copy_funcs here.

Change-Id: I0b727f8fb4e54a6e454277d865e62ef3d5dc6465