Bug #34616560 - Assertion failure: row0sel.cc:4708:prebuilt->sql_stat_start || trx->state.load(std::memory_order_relaxed) == TRX_STATE_ACTIVE
Innodb detects a deadlock with just two transactions,
choosing one as a victim to be rolled back.
This transaction is stopped in the middle of a subquery, but
would continue the query anyways, resulting in a failed
assertion down the line.

This patch forces UpdateRowsIterator::DoImmediateUpdatesAndBufferRowIds
to return return true, thus setting the local_error variable in the reading
loop in UpdateRowsIterator::Read, which stops it from continuing the query.

The query ends with an error message about the deadlock instead.

Note: the main issue is, the code ignored the return value of fill_record,
which can signal an error.

Change-Id: I3278a419d813eb1dcf5c5be1a451436afb2bc7a6