PXB-2878 - Incremental backup prepare displays warning of redo log running out of free space
Problem
-------
A warning is issued during prepare. Apply of incremental on the full backup. This is full prepare.
During the prepare, we see this below warning

`[Warning] [MY-013867] [InnoDB] Redo log is running out of free space, pausing user threads... Consider increasing innodb_redo_log_capacity.`

This occurs because, there is new redo generated by dynamic _metadata(Logical redo logs) before a checkpoint is allowed :-o
With the new design, the existing PXB redo file dressed as ib_redo0 is treated as full and it
cannot accomodate new redo. (it can acommodate a bit due to file extension logic by PXB)

Until now, for rollback, the redo is generated after the checkpoint. So we could recreate the redo files
and this means, we have brand new redo files and accomodate any amount of redo.

For the dynamic metadata, this redo comes before checkpoint. We have to apply this before a new checkpoint is issued.
Now, we cannot resize and recreate the ib_redo0 until a checkpoint is issued.

Fix:
----
The only solution is to accomodate redo generated, and then do a checkpoint.
An additional capacity of 1G is added and this will allow new redo files to be cre
This also allows the older redo files to be recycled.

Once the checkpoint is finished, we will recreate the redo files (exisitng logic)