BUG#34667384 Report flags incorrect for opened event
When opening a NdbDictionary::Event from NDB the report option flags are
not sent to the NdbApi, this causes the Event::getReport()
function to always indicate that only updated columns would be reported
for NdbDictionary::Event's. This problem makes it impossible to use the
NdbApi for checking how an Event is configured in NDB. Furthermore the
Event::getReport() function is documented as returning only one value
while it actually returns a value consisting of several 'EventReport'
flags. Due to this the `ndbinfo.reporting` column is using an ENUM value
which also always shows the same value due to first problem.

Fix by extending the CreateEvntConf signal with an additional data field
to send the report flags and translate them from
CreateEventReq::EventFlags to 'EventReport' flags as used in the NdbApi.
Also change the `ndbinfo.reporting` column to use a SET to describe the
different report flags for each event in NDB, this means that were the
`reporting` column preivosly only returned one value it will now show
more than one. The complete functionality is tested by showing how
events for the well known tables of ndb_ddl
suite are configured.
Add signal data printers for CreateEvntReq, CreateEvntConf and
CreateEvntRef.
Also improve functionality for EventDurability:
 * Change Event::getDurability() to always return correctly that Event
is created as permanent (ED_PERMANENT) which is the only supported
value.
 * Add function comment describing that Event::setDurability() has no
effect, event will alwyas be created as permanent.

Change-Id: I91a47331ffcb74c39b979e8c44c7e1f167ac1c35