Bug #30049013 MEMORY ALLOCATION IMBALANCE IN UNIQUE HASH INDEX
Fix reorganize of unique indexes.
When adding data nodes to an existing cluster new fragments are
mapped to the data nodes but the distribution of unique indexes is
not balanced. The new fragments are spread across all of the nodes
making the existing nodes to have more fragments than the new ones.
This patch changes the Dbdict::create_fragmentation function in
order to make the new fragments to be mapped to new added nodes
making each node have the same share of fragments

Change-Id: I8fa7928f214bc0406fb3b9b2a24f9923c8f32d57