Bug #34640773 	Imbalance in fragment distribution when ClassicFragmentation OFF
When adding data nodes to an existing cluster new fragments are
mapped to the data nodes but the fragment distribution is not
balanced when classicFragmentation is OFF.
The mapping of fragments to nodegroups is correct, but the
'preferred primary' is not balanced across all nodes within
the same nodegroup.
This patch fixes the logPart assigned to logPatIndex during
the execCREATE_FRAGMENTATION_REQ in BDDIH block.

Change-Id: I271157e6dc4c91a6aeccdfdeaec21991ab973c9c