- Fixed pxb-2710-predict_memory.sh test case
Added a condition to only add new 8.0.30+ variable depending on the
version of backup server.
Also fixed a race condition when backup completes before the last
INSERT has the chance to commit. If the INSERT completes in between the
end of backup and before we run the record_db_state we might see
differences on verify_db_state as the last INSERT will be rolled back
by PXB.

 - Fixed all_files.sh test case

We no longer generate redo files to be carried over on copy-back.
Adjusted test case to not check for ib_logfile when running diff