Bug#34644930 NDB: Performance regression in TPC-H Q22
One of the patches for WL#14388, where the particular patch intended
to be a pure refactoring patch without any functional changes, unintentionally
changed Ndb_value::save_in_field() such that a 'type_conversion_status' of
'TYPE_NOTE_TRUNCATED' and 'TYPE_NOTE_TIME_TRUNCATED' were treated as an
error instead of being ignored as it used to be. Thus building of
InterpreterCode for pushed conditions failed, and the condition were
'unpushed' back to the server.

This patch change the handling of such status back to being ignored again
as it used to be prior to:

commit 56740694cb623b00322dcbe20c20db9a7310a3ef
Author: Ole John Aske <ole.john.aske@oracle.com>
Date:   Wed Jun 9 13:10:02 2021 +0200

    WL#14388 Push condition referring fields from an SPJ ancestor table

    Refactoring of class Ndb_item in ha_ndbcluster_cond.cc, no
    functional change.

Change-Id: Icd607cbad0ef64da64c9049bfc7321e82e2a1644