Bug#34377854: Bug that can directly crashes MySQL
If order by references a subquery in select list through an alias,
and if the query block having the order by gets merged, then while deleting
the unused merged derived columns, the subquery gets deleted.
This is not correct as it is still being used in order by. To determine
if a merged derived column is used or not, "derived_used" was being
checked. "derived_used" is not set for this case because resolving for
order by does not make a call to fix_fields() when it is resolved against
alias in the select list.
Since we now have reference counting for items, we can use it to
check if a derived column is used or not. So as part of the solution we
do the following now.
1. Increment reference count when order by references the select
   expression.
2. Check the reference count before removing a merged derived table column.
3. Remove the use of "derived_used".

Change-Id: I65eed6145f27eeaa50206967c2745791c86f4ea2