Bug#34649576 create table with max_rows or partition count may fail after drop nodegroup
Problem:

After drop nodegroup creating (or altering) table using explicit
partition count

  CREATE TABLE t ... ENGINE=NDB PARTITION BY KEY() PARTITIONS 8;

or using max_rows

  CREATE TABLE t ... ENGINE=NDB MAX_ROWS=10000000;

may fail with error:

  Warning 1296    Got error 771 'Given NODEGROUP doesn't exist in this cluster' from NDB
  Error   1005    Can't create table 't5' (use SHOW WARNINGS for more info).

Cause:

Since the above partitioning schemes may create tables there the number
of partitions does not match the nodegroup count the cluster try to put
new partitions to nodegroups in a round robin fashion, partitions for
next created table starts at the nodegroup next to there the last
partition of previous table was put.

If the next nodegroup to use is being dropped one can no longer create
table using the above schemes since the next nodegroup pointer is no
longer valid.

Solution:

If the next nodegroup is dropped, use the nodegroup after that as new
next nodegroup, or if it was the last nodegroup use the first nodegroup
as the new next nodegroup.

If the current next nodegroup is not dropped keep the next nodegroup as
is even if nodegroup index changes.

Change-Id: I615d7d6322df50d64ae37e680ed4d11e272a20a3