BUG#34655310: Make THD constructor explicit
Problem:

The THD::THD(bool) constructor will accept a THD* as argument, since
pointers are implicitly convertible to bool.  Therefore, a function
declared to take a THD& argument will accept a THD*: the pointer is
implicitly converted to a bool, and passed as argument to the
constructor; a new THD is created and passed to the function.  It is
easy to pass a THD* instead of a THD& by mistake (e.g. when a function
is changed from taking a pointer to taking a reference, and one of the
call sites is forgotten).  This would hide the mistake by making it
compile, and later behave in some unexpected way.

Solution:

Declare the constructor as explicit, so that such mistakes produce
compilation errors.

Change-Id: I651a7e07178080357eade6c0a51fbc742e09e86c