Bug#34190122: High Memory Usage When Running Stored Procedure
As part of the window function support for hypergraph, a fix was made to
ensure that a cached item used for execution of window functions points
to the correct item when prepared statements are used. This meant,
registering the change made during each execution so that the next
execution would have the correct item to use.
However, this change was made in Item_cache::store which leads to high
memory usage in case of stored procedures.
Solution is to make this change local to apply_temp_table() where
we already take care of registering changes to  other cached items used
for window functions.

Change-Id: Ia66d73591ea18e375e49e167aec82db3a046d981