Bug #34207811 : Audit plugin does not log multiqueries correctly.
Description :
-------------
The Audit plugin does not log multiqueries correctly. It will always log whole
query instead of the actual query that is being executed. This is because the
query length in the THD is set after the general logging is done.

Fix:
----
Setting the query length of the THD before general logging will ensure that only
one query that is being executed is logged.

Note:
-----
* This patch backports the fix for "Bug #31260698 : PERFORMANCE REGRESSION ON
  WINDOWS INTRODUCED BY FIX FOR BUG#30654405".
* This patch also fixes the "Bug #34569180 : P_S.events_statements_history
  tables store wrong data when executing multiqueries".

Change-Id: I1ec7641599deff284ca06cc0357a61d23d33d90d