Bug#34663681: Rename Item_equal::const_item
Rename Item_equal::const_item to m_const_arg, so that it doesn't
conflict with Item::const_item(). Without this fix, calls to
const_item() on an Item_equal object fail to compile if the Item_equal
object is not cast to Item first.

Also rename the corresponding getter and setter to match the new
name (const_arg() and set_const_arg()).

Change-Id: I44ede67441195a7a9fdf85014e714f7c0dfb3cce