Bug#34603259: Hypergraph : Assertion failure in join_optimizer.cc
When equality propagation simplified a subjunction so that it was
replaced by a conjunction, the conjunction was not split up in
individual predicates which could be pushed down separately. This
confused the condition pushdown logic and could make it fail to push
down conditions correctly.

Fixed by splitting up any Item_cond_and objects that result from
equality propagation.

Change-Id: Ifab122f2c6cbeb3bda16d681784cd4305391e74a