BUG #21498994 mysqlbinlog does not work with start-position larger than 4G
Problem
------------------------------
 mysqlbinlog can read events at positions larger than 4 GB if
 it starts reading from the beginning. It is unable to use a value for
 --start-position that is larger than 4 GB

Analysis/Root-cause analysis
---------------------------------
The problem is with the variable start_position as it's upper limit
is set to 2^32 i.e 4294967854. So on running the commands mysqlbinlog
var/mysqld.1/data/master-bin.000002 --start-position=4294967854,
it gives error. So to read the binary log with greater value,
start_position variable size must be increased to 2^64.
Meanwhile, The option --start-position cannot be used with values
greater than 4 GiB (4294967854), when one of read-from-remote-server,
read-from-remote-source, or read-from-remote-master is used.

Solution
---------------------------------
1. Increasing the start_position variable limit from ~(uint32)0 to ~(uint64)0.
2. When connected to remote server and greater than 4G, generate error.
-------------------------------------------------------------------------------

Change-Id: I7b5b5c111870853dff46d94ed78caf3fe8ef3e02