Bug#34681893 [noclose] testSecureSocket-t failures in PB2
The macro EVP_EC_gen() as defined in OpenSSL 3 may not be usable
in C++ without warnings. The macro uses strstr(curve, "") to cast
curve from a const char * to a char *. In C, strstr() always
returns char *, but C++ strstr() does not work this way.

Define EVP_EC_generate() as a workaround that clearly takes a
const char * argument and does not attempt to cast it.

Also ndb_openssl3_compat.cpp will not build on OpenSSL 1.0.1, so
this patch adds 1.0.2 as a minimum version requirement.

Change-Id: I83de64bc43dcd81a14f7a77f9170dd98ea73d5fa