Bug#34482813: Post-push fix II. Fix spoadic PB2 failures in thread_pool_transaction_delay_debug
Problem: Previous PPF moved checking of debug flags after query
execution. But when returning to TP code after executing a statement, it
is possible that mtr can switch connection and modify the debug flags
before they are read by the TP code, and trigger the assert.

Solution: Sample the debug flag inside the server code after the query
has been read, but before the reply is sent. Store the value in a
thread-local variable which then can be referenced in the TP code.

Additionally, since it is possible that the do-while loop in
tp_process_event() can execute multiple statements, the code to check
transaction delay along with the call into the server to execute a
command/statement is moved to a new function which is called from the
body of the loop. This gives us a convenient location for anything that
needs to be done for each statement/command executed by a TP worker.

Change-Id: Iaf9c6bf8d6d34ae2c125ddcb91ec09169c668c51