Bug#34629157: Erroneous MYSQL_TYPE_INVALID data type
The problem here is in the deduction of data types for dynamic
parameters passed as parameters to a user-defined SQL function.
With a single parameter, the deduction is correct, however with
more than one parameter, no deduction is performed for the second
and following parameters.

The anomaly is due to a typo in function Item_func_sp::fix_fields()
in a loop that processes the parameters. Instead of using a loop
variable, only the first parameter was ever processed.

Test case is made by extending ps_type_deduction with cases for
all data types passed to SQL functions, including having multiple
parameters.

Change-Id: I01a429d2f67435531d91f466935be6dd36a7d447